Hannah_3122225001032_A/requirements.txt
pandas
requests
pymongo
python-dotenv
